+++ CHANGELOG.md
@@ -27,6 +27,8 @@
     within your `plugin` tag for `io.sentry:sentry-maven-plugin`.
     This is not recommended, as using mismatched versions of the Sentry dependencies can introduce build time or run time failures and crashes.
 
+- Upgrade internal Sentry SDK to 8.4.0 ([#134](https://github.com/getsentry/sentry-maven-plugin/pull/134))
+
 ### Dependencies
 
 - Bump CLI from v2.41.1 to v2.42.3 ([#125](https://github.com/getsentry/sentry-maven-plugin/pull/125), [#129](https://github.com/getsentry/sentry-maven-plugin/pull/129), [#132](https://github.com/getsentry/sentry-maven-plugin/pull/132))